Dużo zmian. Wersja całkiem finalna bez integracji z Barobotem.

+++ android/src/com/barobot/gui/dataobjects/Ingredient.java
@@ -14,7 +14,7 @@ public class Ingredient {
 		return what;
 	}
 	
-	public double getQuantity()
+	public int getQuantity()
 	{
 		return quantity;
 	}
